refactor: PreQuestionAlreadyExistException 을 DebugMessage 사용하도록 수정
